wgml: clenaup code
fix a few issues with missing terminate character
do code more portable

+++ bld/wgml1/c/gsfd2c.c
@@ -35,7 +35,7 @@
 /***************************************************************************/
 
 
-condcode    scr_d2c( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_d2c( parm parms[MAX_FUN_PARMS], int parmcount, char **result, int32_t ressize )
 {
     tok_type        parm1;
     condcode        cc;
