+++ bld/wgml1/c/gsfdelst.c
@@ -30,7 +30,7 @@
 /*      &'delstr('abcdef',10) ==> abcdef                                   */
 /***************************************************************************/
 
-condcode    scr_delstr( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_delstr( parm parms[MAX_FUN_PARMS], int parmcount, char **result, int32_t ressize )
 {
     tok_type        parm1;
     condcode        cc;
